Encapsulate Jayson types so they do not spread across the code
