+++ src/lib/http_client.ts
@@ -1,6 +1,8 @@
 import http from 'http';
 import https from 'https';
 import jayson, { HttpsClientOptions } from 'jayson/promise';
+import { HTTPClientInterface, HTTPRequest } from '../types';
+
 // The TCP session constructed by Node HTTP module would get closed after 5 seconds of inactivity by default.
 // Extend this timeout to 30 to reduce the number of sessions constructed.
 const TCP_SESSION_KEEP_ALIVE_MSEC = 30000;
@@ -12,7 +14,26 @@ interface ExtraOptions {
   version?: number
 }
 
-export function constructRPCClient(nodeURL: string, extraOptions: ExtraOptions = {}): jayson.HttpClient | jayson.HttpsClient {
+class JaysonHTTPClient implements HTTPClientInterface {
+  private client: jayson.HttpClient;
+  constructor(_client: jayson.HttpClient) {
+    this.client = _client;
+  }
+  request(method: string, params: any[], id?: string | number): Promise<any> {
+    return this.client.request(method, params, id)
+  }
+
+  generateRequest(method: string, params: any[], id: string | number): HTTPRequest {
+    return {
+      method: "test",
+      params: [],
+      id: ""
+    }
+    //return this.client.request(method, params, id, shouldCall)
+  }
+}
+
+export function constructRPCClient(nodeURL: string, extraOptions: ExtraOptions = {}): HTTPClientInterface {
   const nodeUrl = new URL(nodeURL);
 
   const mergedOptions: HttpsClientOptions = {
@@ -30,11 +51,11 @@ export function constructRPCClient(nodeURL: string, extraOptions: ExtraOptions =
   if (nodeURL.substring(0, 5) === 'https') {
     const agent = new https.Agent(agentOptions);
     mergedOptions.agent = agent;
-    return jayson.client.https(mergedOptions);
+    return new JaysonHTTPClient(jayson.client.https(mergedOptions));
   } else {
     const agent = new http.Agent(agentOptions);
     mergedOptions.agent = agent;
-    return jayson.client.http(mergedOptions);
+    return new JaysonHTTPClient(jayson.client.http(mergedOptions));
   }
 }
