+++ src/types.ts
@@ -7,3 +7,15 @@ export class ExporterPosition {
     this.primaryKey = primaryKey;
   }
 }
+
+export interface HTTPRequest {
+  method: string;
+  params: Array<any>;
+  id: string;
+}
+
+// This interface is modeled agains the jayson Client that we actually use
+export interface HTTPClientInterface {
+  request(method: string, params: any[], id?: string | number): Promise<any>;
+  request(method: string, params: any[], id: string | number, shouldCall: boolean): HTTPRequest;
+}
